Janitor: remove unused code.
