+++ ApplicationLibCode/Commands/CompletionExportCommands/CMakeLists_files.cmake
@@ -19,7 +19,6 @@ set(SOURCE_GROUP_HEADER_FILES
     ${CMAKE_CURRENT_LIST_DIR}/RicWellPathFractureTextReportFeatureImpl.h
     ${CMAKE_CURRENT_LIST_DIR}/RicWellPathFractureReportItem.h
     ${CMAKE_CURRENT_LIST_DIR}/RicExportCompletionsForTemporaryLgrsFeature.h
-    ${CMAKE_CURRENT_LIST_DIR}/RicMswTableFormatterTools.h
     ${CMAKE_CURRENT_LIST_DIR}/RicWellPathExportMswTableData.h
     ${CMAKE_CURRENT_LIST_DIR}/RicMswTableDataTools.h
 )
@@ -45,7 +44,6 @@ set(SOURCE_GROUP_SOURCE_FILES
     ${CMAKE_CURRENT_LIST_DIR}/RicWellPathFractureTextReportFeatureImpl.cpp
     ${CMAKE_CURRENT_LIST_DIR}/RicWellPathFractureReportItem.cpp
     ${CMAKE_CURRENT_LIST_DIR}/RicExportCompletionsForTemporaryLgrsFeature.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/RicMswTableFormatterTools.cpp
     ${CMAKE_CURRENT_LIST_DIR}/RicWellPathExportMswTableData.cpp
     ${CMAKE_CURRENT_LIST_DIR}/RicMswTableDataTools.cpp
 )
