+++ ApplicationLibCode/Commands/CompletionExportCommands/RicMswTableDataTools.cpp
@@ -23,9 +23,6 @@
 #include "CompletionsMsw/RigMswTableData.h"
 #include "RicMswCompletions.h"
 #include "RicMswExportInfo.h"
-#include "RicMswTableFormatterTools.h"
-
-#include "RifTextDataTableFormatter.h"
 
 #include "Well/RigWellPath.h"
 
@@ -185,7 +182,7 @@ void RicMswTableDataTools::collectWelsegsSegment( RigMswTableData&              
     double endMD   = segment->endMD();
 
     std::vector<std::pair<double, double>> segments =
-        RicMswTableFormatterTools::createSubSegmentMDPairs( startMD, endMD, maxSegmentLength, customSegmentIntervals );
+        RicMswTableDataTools::createSubSegmentMDPairs( startMD, endMD, maxSegmentLength, customSegmentIntervals );
 
     CVF_ASSERT( branch->wellPath() );
 
@@ -207,7 +204,7 @@ void RicMswTableDataTools::collectWelsegsSegment( RigMswTableData&              
         double length = 0;
 
         double midPointMD  = 0.5 * ( subStartMD + subEndMD );
-        double midPointTVD = RicMswTableFormatterTools::tvdFromMeasuredDepth( branch->wellPath(), midPointMD );
+        double midPointTVD = RicMswTableDataTools::tvdFromMeasuredDepth( branch->wellPath(), midPointMD );
 
         if ( midPointMD < prevOutMD )
         {
@@ -316,7 +313,7 @@ void RicMswTableDataTools::collectValveWelsegsSegment( RigMswTableData&         
         endMD   = subSegment->endMD();
     }
 
-    auto splitSegments = RicMswTableFormatterTools::createSubSegmentMDPairs( startMD, endMD, maxSegmentLength, customSegmentIntervals );
+    auto splitSegments = RicMswTableDataTools::createSubSegmentMDPairs( startMD, endMD, maxSegmentLength, customSegmentIntervals );
 
     int        outletSegmentNumber = outletSegment ? outletSegment->segmentNumber() : 1;
     const auto linerDiameter       = valve->wellPath()->mswCompletionParameters()->linerDiameter( exportInfo.unitSystem() );
@@ -339,8 +336,8 @@ void RicMswTableDataTools::collectValveWelsegsSegment( RigMswTableData&         
         row.joinSegment = outletSegmentNumber;
         row.branch      = valve->branchNumber();
 
-        const double subStartTVD = RicMswTableFormatterTools::tvdFromMeasuredDepth( valve->wellPath(), subStartMD );
-        const double subEndTVD   = RicMswTableFormatterTools::tvdFromMeasuredDepth( valve->wellPath(), subEndMD );
+        const double subStartTVD = RicMswTableDataTools::tvdFromMeasuredDepth( valve->wellPath(), subStartMD );
+        const double subEndTVD   = RicMswTableDataTools::tvdFromMeasuredDepth( valve->wellPath(), subEndMD );
 
         double depth  = 0;
         double length = 0;
@@ -440,14 +437,14 @@ void RicMswTableDataTools::collectCompletionWelsegsSegments( RigMswTableData&   
         double startTVD = segment->startTVD();
         double endTVD   = segment->endTVD();
 
-        auto splitSegments = RicMswTableFormatterTools::createSubSegmentMDPairs( startMD, endMD, maxSegmentLength, customSegmentIntervals );
+        auto splitSegments = RicMswTableDataTools::createSubSegmentMDPairs( startMD, endMD, maxSegmentLength, customSegmentIntervals );
         for ( const auto& [subStartMD, subEndMD] : splitSegments )
         {
             int subSegmentNumber = ( *segmentNumber )++;
 
             // TODO: Verify this calculation for fractures
-            double subStartTVD = RicMswTableFormatterTools::tvdFromMeasuredDepth( completion->wellPath(), subStartMD );
-            double subEndTVD   = RicMswTableFormatterTools::tvdFromMeasuredDepth( completion->wellPath(), subEndMD );
+            double subStartTVD = RicMswTableDataTools::tvdFromMeasuredDepth( completion->wellPath(), subStartMD );
+            double subEndTVD   = RicMswTableDataTools::tvdFromMeasuredDepth( completion->wellPath(), subEndMD );
 
             if ( completion->completionType() == RigCompletionData::CompletionType::FISHBONES )
             {
@@ -918,4 +915,124 @@ void RicMswTableDataTools::generateWsegSicdTableRecursively( RicMswExportInfo&  
     {
         generateWsegSicdTableRecursively( exportInfo, childBranch, sicdValveData );
     }
+}
+
+//--------------------------------------------------------------------------------------------------
+/// Creates sub-segment MD pairs by combining custom intervals with max segment length subdivision
+/// Custom intervals define exact segment boundaries where specified
+/// Areas without custom intervals use max segment length subdivision (if maxSegmentLength > 0)
+//--------------------------------------------------------------------------------------------------
+std::vector<std::pair<double, double>>
+    RicMswTableDataTools::createSubSegmentMDPairs( double                                        startMD,
+                                                   double                                        endMD,
+                                                   double                                        maxSegmentLength,
+                                                   const std::vector<std::pair<double, double>>& customSegmentIntervals )
+{
+    std::vector<std::pair<double, double>> subSegmentMDPairs;
+
+    // If no custom intervals, use original logic with maxSegmentLength subdivision
+    if ( customSegmentIntervals.empty() || maxSegmentLength <= 0.0 )
+    {
+        int    subSegmentCount  = maxSegmentLength > 0.0 ? (int)( std::trunc( ( endMD - startMD ) / maxSegmentLength ) + 1 ) : 1;
+        double subSegmentLength = ( endMD - startMD ) / subSegmentCount;
+
+        double subStartMD = startMD;
+        double subEndMD   = startMD + subSegmentLength;
+        for ( int i = 0; i < subSegmentCount; ++i )
+        {
+            subSegmentMDPairs.push_back( std::make_pair( subStartMD, subEndMD ) );
+            subStartMD += subSegmentLength;
+            subEndMD = std::min( subEndMD + subSegmentLength, endMD );
+        }
+        return subSegmentMDPairs;
+    }
+
+    // Combine custom intervals with maxSegmentLength subdivision
+    // Collect all boundaries (start, end, custom interval boundaries) and sort them
+    std::set<double> boundaries;
+    boundaries.insert( startMD );
+    boundaries.insert( endMD );
+
+    // Add custom interval boundaries that overlap with [startMD, endMD]
+    for ( const auto& [customStart, customEnd] : customSegmentIntervals )
+    {
+        // Check if custom interval overlaps with [startMD, endMD]
+        if ( customEnd > startMD && customStart < endMD )
+        {
+            // Clip custom interval to [startMD, endMD] range
+            double clippedStart = std::max( customStart, startMD );
+            double clippedEnd   = std::min( customEnd, endMD );
+
+            if ( clippedStart < clippedEnd )
+            {
+                boundaries.insert( clippedStart );
+                boundaries.insert( clippedEnd );
+            }
+        }
+    }
+
+    // Convert boundaries to sorted vector
+    std::vector<double> sortedBoundaries( boundaries.begin(), boundaries.end() );
+
+    // For each gap between boundaries, either:
+    // - Use exact boundary if it's from a custom interval
+    // - Subdivide using maxSegmentLength if it's a gap
+    for ( size_t i = 0; i + 1 < sortedBoundaries.size(); ++i )
+    {
+        double gapStart = sortedBoundaries[i];
+        double gapEnd   = sortedBoundaries[i + 1];
+
+        // Check if this gap is covered by a custom interval
+        bool coveredByCustomInterval = false;
+        for ( const auto& [customStart, customEnd] : customSegmentIntervals )
+        {
+            double clippedStart = std::max( customStart, startMD );
+            double clippedEnd   = std::min( customEnd, endMD );
+
+            // If the gap is fully within a custom interval, use exact boundaries
+            if ( gapStart >= clippedStart && gapEnd <= clippedEnd && std::abs( gapStart - clippedStart ) < 1e-6 &&
+                 std::abs( gapEnd - clippedEnd ) < 1e-6 )
+            {
+                coveredByCustomInterval = true;
+                subSegmentMDPairs.push_back( std::make_pair( gapStart, gapEnd ) );
+                break;
+            }
+        }
+
+        // If not covered by custom interval, subdivide using maxSegmentLength
+        if ( !coveredByCustomInterval && maxSegmentLength > 0.0 )
+        {
+            double gapLength = gapEnd - gapStart;
+            int    subCount  = (int)( std::trunc( gapLength / maxSegmentLength ) + 1 );
+            double subLength = gapLength / subCount;
+
+            double subStart = gapStart;
+            for ( int j = 0; j < subCount; ++j )
+            {
+                double subEnd = ( j == subCount - 1 ) ? gapEnd : subStart + subLength;
+                subSegmentMDPairs.push_back( std::make_pair( subStart, subEnd ) );
+                subStart = subEnd;
+            }
+        }
+        else if ( !coveredByCustomInterval )
+        {
+            // No maxSegmentLength, use gap as-is
+            subSegmentMDPairs.push_back( std::make_pair( gapStart, gapEnd ) );
+        }
+    }
+
+    return subSegmentMDPairs;
+}
+
+//--------------------------------------------------------------------------------------------------
+///
+//--------------------------------------------------------------------------------------------------
+double RicMswTableDataTools::tvdFromMeasuredDepth( gsl::not_null<const RimWellPath*> wellPath, double measuredDepth )
+{
+    auto wellPathGeometry = wellPath->wellPathGeometry();
+    CVF_ASSERT( wellPathGeometry );
+
+    double tvdValue = -wellPathGeometry->interpolatedPointAlongWellPath( measuredDepth ).z();
+
+    return tvdValue;
 }
